Cleaned up and merged Exo modifs
