+++ android/AMP/app/src/main/java/io/streamroot/dna/samples/amp/MainActivity.kt
@@ -3,8 +3,8 @@ package io.streamroot.dna.samples.amp
 import android.content.Intent
 import android.os.Bundle
 import android.widget.AutoCompleteTextView
-import android.widget.Button
 import androidx.appcompat.app.AppCompatActivity
+import kotlinx.android.synthetic.main.activity_main.*
 
 class MainActivity : AppCompatActivity() {
     private lateinit var streamEditText: AutoCompleteTextView
@@ -13,16 +13,17 @@ class MainActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
-        streamEditText = findViewById(R.id.streamEditText)
-
-        findViewById<Button>(R.id.launchButton).setOnClickListener {
-            val intent = Intent(applicationContext, PlayerActivity::class.java)
-
-            intent.putExtra("streamUrl", streamEditText.text.toString())
-            intent.addFlags(Intent.FLAG_ACTIVITY_REORDER_TO_FRONT)
-            intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-
-            applicationContext.startActivity(intent)
+        launchButton.setOnClickListener {
+            PlayerActivity.makeIntent(this,
+                    PlayerActivity.PlayerActivityArgs(
+                            streamEditText.text.toString()
+                    )
+            ).apply {
+                addFlags(Intent.FLAG_ACTIVITY_REORDER_TO_FRONT)
+                addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
+            }.run {
+                startActivity(this)
+            }
         }
     }
 }
