+++ android/AMP/app/src/main/java/io/streamroot/dna/samples/amp/PlayerActivity.kt
@@ -1,69 +1,38 @@
 package io.streamroot.dna.samples.amp
 
 import android.content.Context
-import android.net.Uri
+import android.content.Intent
 import android.os.Bundle
 import android.view.View
 import androidx.appcompat.app.AppCompatActivity
-import io.streamroot.dna.core.DnaClient
-import io.streamroot.dna.utils.stats.StreamStatsManager
 import com.akamai.amp.media.VideoPlayerContainer
 import com.akamai.amp.media.VideoPlayerView
 import android.view.WindowManager
 import com.akamai.amp.media.elements.MediaResource
 import kotlinx.android.synthetic.main.content_player.*
 import com.akamai.amp.media.errors.ErrorType
-import com.akamai.exoplayer2.DefaultLoadControl
-import io.streamroot.dna.core.InformationCallback
-
-class SRModule(
-        context: Context,
-        playerC: VideoPlayerContainer,
-        loadControl: DefaultLoadControl,
-        url: String
-) {
-    private data class DelayedSRConfig(
-        val ampPlayerInteractor: AMPPlayerInteractor,
-        val ampQosModule: AMPQosModule,
-        val ampBandwidthMeter: AMPBandwidthMeter = AMPBandwidthMeter()
-    )
-
-    private var streamStatsManager: StreamStatsManager? = null
-
-    private val dnaDelayedSRConfig = DelayedSRConfig(AMPPlayerInteractor(loadControl), AMPQosModule(playerC))
-    private var dnaClient = runCatching { DnaClient.newBuilder()
-                .context(context)
-                .playerInteractor(dnaDelayedSRConfig.ampPlayerInteractor)
-                .also { AMPSRConfig.configureStreamroot(it) }
-                .qosModule(dnaDelayedSRConfig.ampQosModule)
-                .bandwidthListener(dnaDelayedSRConfig.ampBandwidthMeter)
-                .start(Uri.parse(url))
-    }.getOrNull()
-
-    fun finalUrl() = dnaClient?.manifestUrl
-
-    fun enableStats(callback: InformationCallback)
-            = dnaClient?.let { streamStatsManager = StreamStatsManager.newStatsManager(it, callback) }
-
-    fun setPlayerView(videoPlayerView: VideoPlayerView) {
-        dnaDelayedSRConfig.apply {
-            ampBandwidthMeter.setPlayer(videoPlayerView)
-            ampPlayerInteractor.setPlayer(videoPlayerView)
-        }
-    }
-
-    fun terminate() {
-        dnaClient?.close()
-        dnaClient = null
-
-        streamStatsManager?.close()
-        streamStatsManager = null
-    }
-}
+import io.streamroot.dna.samples.amp.impl.AMPSRModule
 
 class PlayerActivity : AppCompatActivity(), VideoPlayerContainer.VideoPlayerContainerCallback {
 
-    private var srModule: SRModule? = null
+    data class PlayerActivityArgs(val url:String?)
+
+    companion object {
+        private const val ARG_STREAM_URL = "streamUrl"
+
+        fun makeIntent(ctx: Context, args: PlayerActivityArgs) : Intent {
+            return Intent(ctx, PlayerActivity::class.java).apply {
+                putExtra(ARG_STREAM_URL, args.url)
+            }
+        }
+        fun extractArgs(i: Intent) : PlayerActivityArgs {
+            return PlayerActivityArgs(
+                    i.getStringExtra(ARG_STREAM_URL)
+            )
+        }
+    }
+
+    private var srModule: AMPSRModule? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -72,15 +41,15 @@ class PlayerActivity : AppCompatActivity(), VideoPlayerContainer.VideoPlayerCont
 
         videoPlayerContainer.addVideoPlayerContainerCallback(this)
 
-        val url = intent.extras?.getString("streamUrl")?.takeUnless { it.isEmpty() }
-                ?: AMPSRConfig.DEFAULT_VIDEO_URL
+        val args = extractArgs(intent)
+        val url = args.url?.takeUnless { it.isEmpty() } ?: AMPSRConfig.DEFAULT_VIDEO_URL
 
         start(AMPSRConfig.DNA_ENABLED, url)
     }
 
     fun start(srEnabled: Boolean, url: String) = with (if (srEnabled) {
         val loadControl = videoPlayerContainer.setBufferDimensions(5000, 30000, 2500,5000)
-        srModule = SRModule(this, videoPlayerContainer, loadControl, url)
+        srModule = AMPSRModule(this, videoPlayerContainer, loadControl, url)
         srModule?.enableStats(streamrootDnaStatsView)
         srModule?.finalUrl()?.toString() ?: url
     } else {
@@ -111,8 +80,8 @@ class PlayerActivity : AppCompatActivity(), VideoPlayerContainer.VideoPlayerCont
     override fun onResourceError(errorType: ErrorType, exception: Exception) {}
 
     override fun onResume() {
-        videoPlayer()?.onResume()
         super.onResume()
+        videoPlayer()?.onResume()
     }
 
     override fun onPause() {
