+++ android/AMP/app/src/main/java/io/streamroot/dna/samples/amp/impl/AMPBandwidthMeter.kt
@@ -1,4 +1,4 @@
-package io.streamroot.dna.samples.amp
+package io.streamroot.dna.samples.amp.impl
 
 import android.os.Handler
 import com.akamai.amp.media.VideoPlayerView
@@ -8,24 +8,20 @@ import io.streamroot.dna.core.BandwidthListener
 import java.util.concurrent.atomic.AtomicLong
 
 class AMPBandwidthMeter : BandwidthMeter, BandwidthListener {
-    fun setPlayer(player: VideoPlayerView) {
-        player.setBandwidthMeter(this)
-    }
-
-    override fun getTransferListener(): TransferListener? = null
-    override fun addEventListener(
-        eventHandler: Handler?,
-        eventListener: BandwidthMeter.EventListener?
-    ) {}
-    override fun removeEventListener(eventListener: BandwidthMeter.EventListener?) {}
 
     private val estimatedBandwidth = AtomicLong(0L)
 
-    override fun getBitrateEstimate(): Long {
+    fun setPlayer(player: VideoPlayerView) { player.setBandwidthMeter(this) }
+
+    override fun getBitrateEstimate() : Long {
         return estimatedBandwidth.get()
     }
 
     override fun onBandwidthChange(estimatedBandwidth: Long) {
         this.estimatedBandwidth.set(estimatedBandwidth)
     }
+
+    override fun getTransferListener(): TransferListener? = null
+    override fun addEventListener(eventHandler: Handler?, eventListener: BandwidthMeter.EventListener?) {}
+    override fun removeEventListener(eventListener: BandwidthMeter.EventListener?) {}
 }
